pyproject.toml
[tool.poetry]
name = "meh-api"
version = "0.1.0"
description = ""
authors = ["Nathan Evans <nathan@de1ux.com>", "Saint <saintmeh@gmail.com>"]
readme = "README.md"
packages = [{ include = "meh_api" }]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
django-rest-framework = "^0.1.0"
qrcode = "^7.4.2"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
